实现了去年的预测方案（即平动直接用kf；小陀螺用一个ukf滤中心，一个kf滤角度）。 由于是我自己重写的，采用的坐标系和单位都不一样，所以参数还需要重新调。

+++ src/autoaim_prediction/include/pnp_solver.hpp
@@ -74,13 +74,13 @@ public:
     static constexpr float BIG_WIDTH = 0.2253;
     static constexpr float SMALL_WIDTH = 0.135;
     // 装甲板坐标系向右是x，向前是y，向上是z。
-    const std::vector<cv::Point3d> SMALL_POINTS {
+    const std::vector<cv::Point3f> SMALL_POINTS {
         {-SMALL_WIDTH / 2, 0, HEIGHT / 2},
         {-SMALL_WIDTH / 2, 0, -HEIGHT / 2},
         {SMALL_WIDTH / 2, 0, -HEIGHT / 2},
         {SMALL_WIDTH / 2, 0, HEIGHT / 2}
     };
-    const std::vector<cv::Point3d> BIG_POINTS {
+    const std::vector<cv::Point3f> BIG_POINTS {
         {-BIG_WIDTH / 2, 0, HEIGHT / 2},
         {-BIG_WIDTH / 2, 0, -HEIGHT / 2},
         {BIG_WIDTH / 2, 0, -HEIGHT / 2},
